changed nrank() to rank(), now uses nsing parameter to report rank
Ronan Fleming <ronan.mt.fleming@gmail.com> discovered that the rank method was
returning incorrect results for simple matrices.  Turns out that the LUSOL
fortran code is not computing the nrank parameter correctly.  nsing, however,
is correct and is widely using in MINOS and SQOPT.  The rank method now
computes rank as n - nsing.  That is number of columns minus the number of
apparently dependent columns.

I also changed the name of the method from nrank to rank.  This overloads
matlab's rank method and looks nicer.